docs: .gitignore and remove Debug files

+++ Hardware/Serial/Serial.h
@@ -4,7 +4,7 @@
 #include "ti_msp_dl_config.h"
 
 // 定义接收数据变量
-extern uint8_t uart_data;
+extern int uart_data;
 
 void uart0_send_char(char ch);
 void uart0_send_string(char* str);
